ENH: Fixed registration.
In order to deal with the initialization of center of rotation,
the new registration framework must use the initial fixed transform
and the initial moving transforms, to move the origin to and from
the center of rotation.

Additionally, when computing the final resampling, a CompositeTransform
must be used to compose the final Fixed transform with the transform
computed from the registration and the initial Moving Transform.